remove old metric configuration, add blue icon if at least one definition selected, some refactoring
