+++ publisher/src/components/Menu/Menu.tsx
@@ -69,7 +69,6 @@ const Menu: React.FC = () => {
   const location = useLocation();
   const windowWidth = useWindowWidth();
   const [settingsSearchParams] = useSettingsSearchParams();
-  const { system: systemSearchParam } = settingsSearchParams;
   const [isMobileMenuOpen, setIsMobileMenuOpen] = useState(false);
 
   const pathWithoutAgency = removeAgencyFromPath(location.pathname);
@@ -224,7 +223,7 @@ const Menu: React.FC = () => {
               }}
             >
               Get Started
-              {/* Guidance: Metric Configuration Progress Toast */}
+              {/* Guidance: Metric MetricsConfiguration Progress Toast */}
               {isMetricConfigStep && hasSystemMetricParams && (
                 <ProgressTooltipToast showToast={showMetricConfigProgressToast}>
                   {metricConfigurationProgressSteps.map((step) => {
@@ -284,6 +283,17 @@ const Menu: React.FC = () => {
           Data
         </MenuItem>
 
+        {/* Metric Config */}
+        <MenuItem
+          onClick={() => {
+            navigate("metric-config");
+            handleCloseMobileMenu();
+          }}
+          active={pathWithoutAgency === "metric-config"}
+        >
+          Configuration
+        </MenuItem>
+
         {/* Learn More */}
         {windowWidth > MIN_TABLET_WIDTH && (
           <MenuItem>
@@ -329,15 +339,7 @@ const Menu: React.FC = () => {
               <SubMenuItem
                 key={path}
                 onClick={() => {
-                  if (path === "metric-config") {
-                    navigate(
-                      systemSearchParam
-                        ? `settings/${path}?system=${systemSearchParam}`
-                        : `settings/${path}`
-                    );
-                  } else {
-                    navigate(`settings/${path}`);
-                  }
+                  navigate(`settings/${path}`);
                   handleCloseMobileMenu();
                 }}
               >
