--- publisher/src/components/MetricConfiguration/RaceEthnicitiesForm.tsx
@@ -1,294 +1,0 @@
-// Recidiviz - a data platform for criminal justice reform
-// Copyright (C) 2023 Recidiviz, Inc.
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with this program.  If not, see <https://www.gnu.org/licenses/>.
-// =============================================================================
-
-import { debounce } from "lodash";
-import { observer } from "mobx-react-lite";
-import React, { useRef } from "react";
-import { useParams } from "react-router-dom";
-
-import { useStore } from "../../stores";
-import { BinaryRadioButton } from "../Forms";
-import { getActiveSystemMetricKey, useSettingsSearchParams } from "../Settings";
-import {
-  Ethnicity,
-  Header,
-  RaceContainer,
-  RaceDisplayName,
-  RaceEthnicitiesContainer,
-  RaceEthnicitiesDescription,
-  RaceEthnicitiesDisplay,
-  RaceEthnicitiesTitle,
-  raceEthnicityGridStates,
-  RaceEthnicityRadioButtonGroupWrapper,
-  RaceSelection,
-  RaceWrapper,
-  RadioButtonGroupWrapper,
-  sortRaces,
-  SpecifyEthnicityWrapper,
-  StateKeys,
-  Subheader,
-} from ".";
-
-export const RACE_ETHNICITIES_DESCRIPTION =
-  "This breakdown asks for combinations of race and ethnicity, and should be based on what data is available via your case management system. Answering all of the questions below will fill out the grid for this breakdown.";
-
-export const RaceEthnicitiesForm = observer(() => {
-  const { agencyId } = useParams() as { agencyId: string };
-  const [settingsSearchParams] = useSettingsSearchParams();
-  const { metricConfigStore } = useStore();
-  const {
-    metrics,
-    getEthnicitiesByRace,
-    updateAllRaceEthnicitiesToDefaultState,
-    updateRaceDimensions,
-    saveMetricSettings,
-  } = metricConfigStore;
-
-  const { system: systemSearchParam, metric: metricSearchParam } =
-    settingsSearchParams;
-
-  const systemMetricKey = getActiveSystemMetricKey(settingsSearchParams);
-
-  const ethnicitiesByRace =
-    (systemSearchParam &&
-      metricSearchParam &&
-      getEthnicitiesByRace(systemSearchParam, metricSearchParam)) ||
-    {};
-
-  const ethnicitiesByRaceArray = Object.entries(ethnicitiesByRace);
-
-  const canSpecifyEthnicity =
-    ethnicitiesByRaceArray.filter(([_, ethnicities]) => {
-      /** At least one Race where all three Ethnicities are enabled */
-      return (
-        ethnicities[Ethnicity.HISPANIC_OR_LATINO].enabled &&
-        ethnicities[Ethnicity.NOT_HISPANIC_OR_LATINO].enabled &&
-        ethnicities[Ethnicity.UNKNOWN_ETHNICITY].enabled
-      );
-    }).length > 0;
-
-  const specifiesHispanicAsRace =
-    ethnicitiesByRaceArray.filter(([race, ethnicities]) => {
-      /** Unknown Race has both Hispanic and Not Hispanic enabled */
-      return (
-        race === "Unknown" &&
-        ethnicities[Ethnicity.HISPANIC_OR_LATINO].enabled &&
-        ethnicities[Ethnicity.NOT_HISPANIC_OR_LATINO].enabled
-      );
-    }).length > 0;
-
-  const determineCurrentState = () => {
-    if (canSpecifyEthnicity) {
-      return "CAN_SPECIFY_ETHNICITY";
-    }
-    if (specifiesHispanicAsRace) {
-      return "NO_ETHNICITY_HISPANIC_AS_RACE";
-    }
-    return "NO_ETHNICITY_HISPANIC_NOT_SPECIFIED";
-  };
-
-  const currentState = determineCurrentState();
-
-  const debouncedSave = useRef(debounce(saveMetricSettings, 1000)).current;
-
-  const handleUpdateAllRaceEthnicitiesToDefaultState = (state: StateKeys) => {
-    if (!systemSearchParam || !metricSearchParam) return;
-    const updatedDimensions = updateAllRaceEthnicitiesToDefaultState(
-      state,
-      raceEthnicityGridStates,
-      systemSearchParam,
-      metricSearchParam
-    );
-    debouncedSave(updatedDimensions, agencyId);
-  };
-
-  const handleUpdateRaceDimensions = (race: string, status: boolean) => {
-    if (!systemSearchParam || !metricSearchParam) return;
-    /**
-     * When Unknown Race is disabled in NO_ETHNICITY_HISPANIC_AS_RACE state, we automatically switch
-     * to the NO_ETHNICITY_HISPANIC_NOT_SPECIFIED state because the Unknown Race (Hispanic/Latino Ethnicity)
-     * dimension is the only dimension an agency can specify their numbers for Hispanic/Latino as a Race (while
-     * in the NO_ETHNICITY_HISPANIC_AS_RACE state).
-     */
-    if (
-      !status &&
-      race === "Unknown" &&
-      currentState === "NO_ETHNICITY_HISPANIC_AS_RACE"
-    ) {
-      updateAllRaceEthnicitiesToDefaultState(
-        "NO_ETHNICITY_HISPANIC_NOT_SPECIFIED",
-        raceEthnicityGridStates,
-        systemSearchParam,
-        metricSearchParam
-      );
-    }
-
-    const updatedDimensions = updateRaceDimensions(
-      race,
-      status,
-      currentState,
-      raceEthnicityGridStates,
-      systemSearchParam,
-      metricSearchParam
-    );
-    debouncedSave(updatedDimensions, agencyId);
-  };
-
-  return (
-    <RaceEthnicitiesContainer>
-      <RaceEthnicitiesDisplay enabled={metrics[systemMetricKey]?.enabled}>
-        <RaceEthnicitiesTitle>Race and Ethnicity</RaceEthnicitiesTitle>
-
-        <RaceEthnicitiesDescription>
-          {RACE_ETHNICITIES_DESCRIPTION}
-        </RaceEthnicitiesDescription>
-
-        <SpecifyEthnicityWrapper>
-          <Header>
-            Are you able to record an individual’s <strong>ethnicity</strong> (
-            {Ethnicity.HISPANIC_OR_LATINO}, {Ethnicity.NOT_HISPANIC_OR_LATINO},
-            or {Ethnicity.UNKNOWN_ETHNICITY}) separately from their race in your
-            case management system?
-          </Header>
-
-          <RadioButtonGroupWrapper>
-            <BinaryRadioButton
-              type="radio"
-              id="specify-ethnicity-yes"
-              name="specify-ethnicity"
-              label="Yes"
-              value="yes"
-              checked={canSpecifyEthnicity}
-              onChange={() =>
-                handleUpdateAllRaceEthnicitiesToDefaultState(
-                  "CAN_SPECIFY_ETHNICITY"
-                )
-              }
-            />
-            <BinaryRadioButton
-              type="radio"
-              id="specify-ethnicity-no"
-              name="specify-ethnicity"
-              label="No"
-              value="no"
-              checked={!canSpecifyEthnicity}
-              onChange={() =>
-                handleUpdateAllRaceEthnicitiesToDefaultState(
-                  "NO_ETHNICITY_HISPANIC_AS_RACE"
-                )
-              }
-            />
-          </RadioButtonGroupWrapper>
-        </SpecifyEthnicityWrapper>
-
-        <Header>
-          Which of the following categories does your case management system
-          capture for race?
-        </Header>
-        <Subheader>
-          Fill out a response for each of the following race categories.
-        </Subheader>
-
-        <RaceContainer>
-          {/* Hispanic/Latino as Race (if user cannot specify ethnicity) */}
-          {!canSpecifyEthnicity && (
-            <RaceWrapper>
-              <RaceDisplayName>{Ethnicity.HISPANIC_OR_LATINO}</RaceDisplayName>
-              <RaceSelection>
-                <RaceEthnicityRadioButtonGroupWrapper>
-                  <BinaryRadioButton
-                    type="radio"
-                    id="hispanic-latino-no"
-                    name="hispanic-latino"
-                    label="No"
-                    value="no"
-                    checked={!specifiesHispanicAsRace}
-                    lastOptionBlue
-                    buttonSize="small"
-                    onChange={() =>
-                      handleUpdateAllRaceEthnicitiesToDefaultState(
-                        "NO_ETHNICITY_HISPANIC_NOT_SPECIFIED"
-                      )
-                    }
-                  />
-                  <BinaryRadioButton
-                    type="radio"
-                    id="hispanic-latino-yes"
-                    name="hispanic-latino"
-                    label="Yes"
-                    value="yes"
-                    checked={specifiesHispanicAsRace}
-                    lastOptionBlue
-                    buttonSize="small"
-                    onChange={() =>
-                      handleUpdateAllRaceEthnicitiesToDefaultState(
-                        "NO_ETHNICITY_HISPANIC_AS_RACE"
-                      )
-                    }
-                  />
-                </RaceEthnicityRadioButtonGroupWrapper>
-              </RaceSelection>
-            </RaceWrapper>
-          )}
-
-          {/* Races (Enable/Disable) */}
-          {Object.entries(ethnicitiesByRace)
-            .sort(sortRaces)
-            .map(([race, ethnicities]) => {
-              const raceEnabled = Boolean(
-                Object.values(ethnicities).find(
-                  (ethnicity) => ethnicity.enabled
-                )
-              );
-
-              return (
-                <RaceWrapper key={race}>
-                  <RaceDisplayName>{race}</RaceDisplayName>
-                  <RaceSelection>
-                    <RaceEthnicityRadioButtonGroupWrapper>
-                      <BinaryRadioButton
-                        type="radio"
-                        id={`${race}-no`}
-                        name={`${race}`}
-                        label="No"
-                        value="no"
-                        checked={!raceEnabled}
-                        lastOptionBlue
-                        buttonSize="small"
-                        onChange={() => handleUpdateRaceDimensions(race, false)}
-                      />
-                      <BinaryRadioButton
-                        type="radio"
-                        id={`${race}-yes`}
-                        name={`${race}`}
-                        label="Yes"
-                        value="yes"
-                        checked={raceEnabled}
-                        lastOptionBlue
-                        buttonSize="small"
-                        onChange={() => handleUpdateRaceDimensions(race, true)}
-                      />
-                    </RaceEthnicityRadioButtonGroupWrapper>
-                  </RaceSelection>
-                </RaceWrapper>
-              );
-            })}
-        </RaceContainer>
-      </RaceEthnicitiesDisplay>
-    </RaceEthnicitiesContainer>
-  );
-});
