+++ publisher/src/components/Settings/SettingsMenu.tsx
@@ -16,87 +16,30 @@
 // =============================================================================
 
 import { observer } from "mobx-react-lite";
-import React, { Fragment } from "react";
+import React from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 
 import { settingsMenuPaths } from "../../pages/Settings";
-import { useStore } from "../../stores";
 import { removeAgencyFromPath } from "../../utils";
-import {
-  MenuItem,
-  SettingsMenuContainer,
-  SubMenuListContainer,
-  SubMenuListItem,
-  useSettingsSearchParams,
-} from ".";
+import { MenuItem, SettingsMenuContainer } from ".";
 
 export const SettingsMenu: React.FC = observer(() => {
-  const [settingsSearchParams, setSettingsSearchParams] =
-    useSettingsSearchParams();
   const location = useLocation();
   const navigate = useNavigate();
-  const { metricConfigStore } = useStore();
-  const { getMetricsBySystem } = metricConfigStore;
-
-  const { system: systemSearchParam, metric: metricSearchParam } =
-    settingsSearchParams;
-
-  const handleMetricListItemClick = (metricKey: string) => {
-    setSettingsSearchParams(
-      {
-        system: systemSearchParam,
-        metric: metricKey,
-      },
-      true
-    );
-  };
 
   return (
     <SettingsMenuContainer>
       {/* TODO remove that when team management is finished */}
       {settingsMenuPaths.map(({ displayLabel, path }) => (
-        <Fragment key={path}>
-          <MenuItem
-            selected={
-              removeAgencyFromPath(location.pathname) === `settings/${path}`
-            }
-            onClick={() => {
-              if (path === "metric-config") {
-                return navigate(
-                  systemSearchParam
-                    ? `${path}?system=${systemSearchParam}`
-                    : path
-                );
-              }
-              navigate(path);
-            }}
-          >
-            {displayLabel}
-          </MenuItem>
-
-          {/* Metrics Navigation (appears when a metric has been 
-              selected and allows users to toggle between metrics) */}
-          {removeAgencyFromPath(location.pathname) ===
-            "settings/metric-config" &&
-            path === "metric-config" &&
-            metricSearchParam && (
-              <SubMenuListContainer>
-                {getMetricsBySystem(systemSearchParam)?.map(
-                  ({ key, metric }) => {
-                    return (
-                      <SubMenuListItem
-                        key={key}
-                        activeSection={key === metricSearchParam}
-                        onClick={() => handleMetricListItemClick(key)}
-                      >
-                        {metric.label}
-                      </SubMenuListItem>
-                    );
-                  }
-                )}
-              </SubMenuListContainer>
-            )}
-        </Fragment>
+        <MenuItem
+          key={path}
+          selected={
+            removeAgencyFromPath(location.pathname) === `settings/${path}`
+          }
+          onClick={() => navigate(path)}
+        >
+          {displayLabel}
+        </MenuItem>
       ))}
     </SettingsMenuContainer>
   );
